chore(gateway): move to gateway
