+++ packages/administration/src/gateway/gateway.module.ts
@@ -6,7 +6,7 @@ import { NoDataModule } from '@xm-ngx/components/no-data';
 import { XmPermissionModule } from '@xm-ngx/core/permission';
 import { XmTranslationModule } from '@xm-ngx/translation';
 import { NgJhipsterModule } from 'ng-jhipster';
-import { XmSharedModule } from '../../shared/shared.module';
+import { XmSharedModule } from '../../../../src/app/shared/shared.module';
 import { JhiGatewayComponent } from './gateway.component';
 
 @NgModule({
