+++ packages/administration/src/gateway/index.ts
@@ -1,0 +1,2 @@
+export { GatewayModule } from './gateway.module';
+export { JhiGatewayComponent } from './gateway.component';
