--- src/app/admin/gateway/index.ts
@@ -1,1 +1,0 @@
-export { GatewayModule } from './gateway.module';
